Refs #22285 - Prepare tests for strong params enforcement
- Fix broken hosts api tests
- Fix role cloning tests
- Fix ssh key controller api test
- Fix http proxies controller test
- Fix lookup key override api controller test
- Fix puppetclass controller api test
- fix operatingsystem controller api test
- Fix hostgroups controller test
- Fix integration tests with template fields
  We remove the template fields using JS, so we have to test these forms
  with JS or they fail because the template fields are sent
- Fix host instegration tests
- fix hosts controller tests
- Fix images controller parameter handling
- Fix interface controller test
- Fix puppetclass controller tests
- Fix variable LK controller test
