+++ app/controllers/images_controller.rb
@@ -48,6 +48,6 @@ class ImagesController < ApplicationController
   private
 
   def find_compute_resource
-    @compute_resource = ComputeResource.authorized(:view_compute_resources).find(params[:compute_resource_id])
+    @compute_resource = ComputeResource.authorized(:view_compute_resources).find(params.delete(:compute_resource_id))
   end
 end
