+++ test/controllers/api/v2/http_proxies_controller_test.rb
@@ -26,7 +26,7 @@ module Api
 
       def test_create
         name = 'http_proxy_is_smart'
-        post :create, params: { :http_proxy => { :name => name, :url => 'http://what????', :port => 5000 } }, session: set_session_user
+        post :create, params: { :http_proxy => { :name => name, :url => 'http://what????:5000' } }, session: set_session_user
 
         assert_response :success
         assert HttpProxy.find_by(:name => name)
