+++ test/integration/puppetclass_lookup_key_js_test.rb
@@ -56,4 +56,15 @@ class PuppetclassLookupKeyJSTest < IntegrationTestWithJavascript
 
     assert page.find("#puppetclass_lookup_key_hidden_value").checked?
   end
+
+  test "edit page" do
+    visit puppetclass_lookup_keys_path
+    within(:xpath, "//table") do
+      click_link "ssl"
+    end
+    fill_in "puppetclass_lookup_key_description", :with => "test"
+    fill_in "puppetclass_lookup_key_default_value", :with => "false"
+    assert_submit_button(puppetclass_lookup_keys_path)
+    assert page.has_link? 'ssl'
+  end
 end
