+++ test/integration/variable_lookup_key_js_test.rb
@@ -27,4 +27,25 @@ class VariableLookupKeyJSTest < IntegrationTestWithJavascript
     value_selector = page.first(".lookup_values table tbody tr textarea")
     assert_equal "", value_selector.value
   end
+
+  test "edit page" do
+    visit variable_lookup_keys_path
+    within(:xpath, "//table") do
+      click_link "special_info"
+    end
+    fill_in "variable_lookup_key_key", :with => "webport"
+    select2 'base', :from => 'variable_lookup_key_puppetclass_id'
+    fill_in "variable_lookup_key_default_value", :with => "8080"
+    assert_submit_button(variable_lookup_keys_path)
+    assert page.has_link? 'webport'
+  end
+
+  test "create new page" do
+    visit variable_lookup_keys_path
+    first(:link, "Create Smart Variable").click
+    fill_in "variable_lookup_key_key", :with => "test"
+    fill_in "variable_lookup_key_default_value", :with => "test"
+    assert_submit_button(variable_lookup_keys_path)
+    assert page.has_link? 'test'
+  end
 end
